Brings in all work from the fb-terrm framebuffer terminal session:

- kernel_entry.S: fix SysV ABI stack alignment (sub rax, 8 before jmp)
- kernel_main.c: remove redundant kmalloc stack switch
- framebuffer.h/c: FB_ANSI_PALETTE visibility("hidden") — fixes GOT/PE
  crash in apply_sgr during EXEC-DOE with framebuffer active
- tools/fbtest.c: host render harness + PNG output + 8 pixel assertions
- scripts/qemu_screenshot.sh: QEMU HMP screendump → PNG
- Makefile.starkernel: fbtest + screendump wired into amd64 qemu target
- docs/lithosananke/kernel-args/DESIGN.md: KernelArgs design (JVM-style
  UEFI command-line overrides for stack, heap, --doe, --log-level)
- docs/NEXT_SESSION.md: task queue for next session
- CLAUDE.md: one-process-at-a-time build rule

EXEC-DOE now runs to full completion (160,673 rows) without GP fault.
amd64 acceptance confirmed clean.

+++ src/inference_engine.c
@@ -412,7 +412,7 @@ uint32_t find_variance_inflection(
         }
 
         for (uint32_t i = 0; i < num_chunks; i++) {
-            uint64_t chunk_start = i * size;
+            uint64_t chunk_start = (uint64_t)i * (uint64_t)size;
             chunk_vars[i] = compute_variance_q48(
                 &heat_data[chunk_start],
                 size
